+++ src/components/Common/CloseIcon.jsx
@@ -1,0 +1,20 @@
+const CloseIcon = ({ width = 16, height = 16 }) => (
+    <svg
+        width={width}
+        height={height}
+        viewBox="0 0 16 16"
+        fill="none"
+        xmlns="http://www.w3.org/2000/svg"
+    >
+        <path
+            d="M0.414336 14.1421L14.5565 0L15.9707 1.41421L1.82855 15.5563L0.414336 14.1421Z"
+            fill="#000"
+        />
+        <path
+            d="M1.41421 0.142113L15.5563 14.2842L14.1421 15.6985L0 1.55633L1.41421 0.142113Z"
+            fill="#000"
+        />
+    </svg>
+);
+
+export default CloseIcon;
