+++ src/components/index.js
@@ -1,15 +1,20 @@
 // Grids
-import SpaceBar from "./Grid/SpaceBar";
-import Container from "./Grid/Container";
+import SpaceBar from "./grid/SpaceBar";
+import Container from "./grid/Container";
 export { SpaceBar, Container };
 
 // Common
-import Breadcrumb from "./Common/Breadcrumb";
-export { Breadcrumb };
+import CloseIcon from "./common/CloseIcon";
+import Breadcrumb from "./common/Breadcrumb";
+export { Breadcrumb, CloseIcon };
 
-// Products
-import ProductItem from "./ProductsItem";
-export { ProductItem };
+// Shop
+import Cart from "./shop/Cart";
+import CartItem from "./shop/CartItem";
+import CartToggleButton from "./shop/CartToggleButton";
+import ProductItem from "./shop/ProductItem";
+import QuantityButton from "./shop/QuantityButton";
+export { ProductItem, QuantityButton, Cart, CartItem, CartToggleButton };
 
 // Other
 import GitHubButton from "./other/GitHubButton";
