+++ src/components/shop/Cart.jsx
@@ -1,5 +1,6 @@
 import React, { useEffect, useMemo, useRef } from "react";
 import { useDispatch, useSelector } from "react-redux";
+import { CloseIcon, CartItem } from "../index";
 import {
     removeItem,
     updateQuantity,
@@ -8,104 +9,6 @@ import {
     closeCart,
 } from "../../features/cart/cartSlice";
 import { formatPrice } from "../../utils/formatPrice";
-
-const CloseIcon = ({ width = 16, height = 16 }) => (
-    <svg
-        width={width}
-        height={height}
-        viewBox="0 0 16 16"
-        fill="none"
-        xmlns="http://www.w3.org/2000/svg"
-    >
-        <path
-            d="M0.414336 14.1421L14.5565 0L15.9707 1.41421L1.82855 15.5563L0.414336 14.1421Z"
-            fill="#000"
-        />
-        <path
-            d="M1.41421 0.142113L15.5563 14.2842L14.1421 15.6985L0 1.55633L1.41421 0.142113Z"
-            fill="#000"
-        />
-    </svg>
-);
-
-const QuantityButton = ({
-    quantity,
-    onIncrement,
-    onDecrement,
-    disabledDecrement,
-}) => (
-    <div className="gap-1 flex items-center">
-        <button
-            className="w-6 h-6 flex items-center justify-center cursor-pointer"
-            onClick={onDecrement}
-            disabled={disabledDecrement}
-        >
-            -
-        </button>
-        <span>{quantity}</span>
-        <button
-            className="w-6 h-6 flex items-center justify-center cursor-pointer"
-            onClick={onIncrement}
-        >
-            +
-        </button>
-    </div>
-);
-
-const CartItem = ({ item, onRemove, onQuantityChange }) => (
-    <li className="[&:not(:last-child)]:border-b-[1px] border-solid border-gray-200 pb-4">
-        <div className="relative">
-            <div className="gap-4 flex justify-between">
-                <div>
-                    <img
-                        src={item.image}
-                        alt={item.name}
-                        className="w-[90px] h-[90px] object-cover"
-                    />
-                </div>
-                <div className="w-full flex flex-col gap-2">
-                    <div className="flex flex-col pt-2">
-                        <h6>{item.name}</h6>
-                        <ul className="gap-1 flex items-center text-xs">
-                            <li>
-                                <span>Size: </span>
-                                <span>XL</span>,
-                            </li>
-                            <li>
-                                <span>Color: </span>
-                                <span>Blue</span>
-                            </li>
-                        </ul>
-                        <button
-                            className="top-2 right-0 w-4 h-4 flex items-center justify-center absolute cursor-pointer"
-                            onClick={() => onRemove(item.id)}
-                        >
-                            <CloseIcon width={8} height={8} />
-                        </button>
-                    </div>
-                    <div className="gap-4 flex items-center justify-between">
-                        <QuantityButton
-                            quantity={item.quantity}
-                            onIncrement={() =>
-                                onQuantityChange(item.id, item.quantity + 1)
-                            }
-                            onDecrement={() =>
-                                onQuantityChange(item.id, item.quantity - 1)
-                            }
-                            disabledDecrement={item.quantity <= 1}
-                        />
-                        <div>
-                            <span>$</span>
-                            <span className="text-lg">
-                                {formatPrice(item.price)}
-                            </span>
-                        </div>
-                    </div>
-                </div>
-            </div>
-        </div>
-    </li>
-);
 
 const Cart = () => {
     const dispatch = useDispatch();
